fix(build): resolve type mismatches and injection errors in AI features
- Fixed: Type mismatch in  by explicitly handling Double/Int conversions.
- Fixed: Removed  annotation from  properties as Views are not Hilt-injected.
- Fixed: Resolved missing imports for  and .
- Verified: Settings injection now correctly relies on the  method.

+++ app/src/main/kotlin/org/koitharu/kotatsu/reader/ui/ai/AiFeatureManager.kt
@@ -360,8 +360,8 @@ class AiFeatureManager @Inject constructor(
 			if (centerX !in 0 until width || centerY !in 0 until height) return textRect
 
 			// Reduced max expansion to prevent merging separate bubbles
-			val maxExpandX = (textRect.width() * 1.5).coerceAtMost((width * 0.25f).toInt()).coerceAtLeast(50)
-			val maxExpandY = (textRect.height() * 1.5).coerceAtMost((height * 0.25f).toInt()).coerceAtLeast(50)
+			val maxExpandX = (textRect.width().toDouble() * 1.5).coerceAtMost((width * 0.25).toDouble()).coerceAtLeast(50.0).toInt()
+			val maxExpandY = (textRect.height().toDouble() * 1.5).coerceAtMost((height * 0.25).toDouble()).coerceAtLeast(50.0).toInt()
 
 			var left = textRect.left
 			var dist = 0
